seccomp: ignore secure_computing return values
This change is inspired by
  https://lkml.org/lkml/2012/4/16/14
which fixes the build warnings for arches that don't support
CONFIG_HAVE_ARCH_SECCOMP_FILTER.

In particular, there is no requirement for the return value of
secure_computing() to be checked unless the architecture supports
seccomp filter.  Instead of silencing the warnings with (void)
a new static inline is added to encode the expected behavior
in a compiler and human friendly way.

v2: - cleans things up with a static inline
    - removes sfr's signed-off-by since it is a different approach
v1: - matches sfr's original change

Reported-by: Stephen Rothwell <sfr@canb.auug.org.au>
Signed-off-by: Will Drewry <wad@chromium.org>
Acked-by: Kees Cook <keescook@chromium.org>
Signed-off-by: James Morris <james.l.morris@oracle.com>

Conflicts:
	arch/microblaze/kernel/ptrace.c
	arch/mips/kernel/ptrace.c
	arch/powerpc/kernel/ptrace.c
	arch/s390/kernel/ptrace.c
	arch/sh/kernel/ptrace_32.c
	arch/sh/kernel/ptrace_64.c
	arch/sparc/kernel/ptrace_64.c

Change-Id: I21b1451dcdff50cc6d47c5b6209d3b5f13c24d25